Move get_full_table_name function
